profiles: mask ~x11-drivers/nvidia-drivers-580.105.08
I'm affected as well but originally underestimated the issue because
I thought it was related to something custom I'm doing, and thought it
may be Xorg-only on top.

Seeing the increasing amount of reports on NVIDIA forums and others, I
think this is better off masked. NVIDIA has acknowledged the issue and
treating it as high priority but it tend to take time before a new
version still.

Like noted in the mask messages, users without issues should feel
welcome to unmask. This by no means affect everyone, so far only
reports come from DP or HDMI users (possibly only HDMIv1) -- and even
then it may be fine for such users depending on the card or such.Also
a non-issue for users actually using 1920x1080@60.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>