dev-qt/qtgraphs: skip hanging tst_qgqmltest in live
Just noticed that tests weren't being ran in 6.10.0, but Qt fixed
something (not an ebuild issue) and now they're being ran which
revealed that this test hangs and ultimately timeouts with
QT_QPA_PLATFORM=offscreen. Tried it on its own with =xcb and it
passes fine, so just going to skip it.

May or may not only happen with USE=quick3d given the warnings
but haven't verified.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>