dev-util/qbs: add 3.1.1
Been procrastinating on bumping this, but tried now given ran into
a test build failure with (upcoming) Qt 6.10.1. Unfortunately did
not fix it, nor has a fix to allow dropping qt610.patch. Given
the test that fail to build is a test we were skipping either way,
just going to skip building it on top for now rather than spend
time on this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>