sci-libs/vtk: increase memory requirement for cuda compile jobs
This reduces the pool size for compilation. The value for CHECKREQS_MEMORY stays
the same as there seem to be only two jobs with 14 GiB usage and the rest stay
below that. So for machines with more memory (>32 GiB) this works as well.

Closes: https://bugs.gentoo.org/932464
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44665
Closes: https://github.com/gentoo/gentoo/pull/44665
Signed-off-by: Sam James <sam@gentoo.org>