app-emulation/vkd3d-proton: update live
Submodules are getting a bit of a pain to maintain, there's now
an extra layer that need a (3rd) copy of spirv headers with its
own mismatching hash. Probably time to give this up, allow all
submodules for live, and do our own tarball for releases.

As a reminder, unfortunately, upstream's vkd3d-proton-3.0.tar.zst
is the prebuilts and not the sources.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>