app-text/mdbook: add 0.5.0
This is a new major version with breaking changes that may require
migration and/or ebuild adjusting given it can create output files
in a different directory depending on how it is used.

Could've considered adding it masked/unkeyworded (or *gasp*, slotted)
but, given there is only one revdep, it can be easily adjusted.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>