dev-util/maturin: adjust for mdbook-0.5
Font licenses vary depending on which version of mdbook docs were
generated with so, to avoid ambiguity and keep it simple, set bounds.
No need for revbumps given the old licenses remain correct until it
is rebuilt (not that removal of CC-BY-4.0 would really be worth a
revbump either way).

`-d html` now creates files in the current directory rather than
the book's (guide/html) and, rather than use that, may as well
change it to "${T}"/html. Haven't tested if using an absolute path
works with the old version, but won't have to worry about this thanks
to the lower bound either way.

maturin's book has a different look with 0.5 which may or may not be
right/desirable, but it is entirely usable so not going to worry too
much about this. I'd assume upstream will update their usage later
which may hard-require >=0.5 either way.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>