dev-qt/qtbase: warn about USE=custom-cflags if have rdrnd
Wouldn't hurt given those that have that enabled and had it working
fine may not realize that it's related. For new installs it's less
of an issue though, so don't think need to keep this forever.

Alternatively could workaround, but I don't want to play in that
film again and would rather encourage users to turn this off unless
they really need it.

As a reminder, for Qt, USE=-custom-cflags does *not* run strip-flags
except for qtwebengine. It's a different form of modifications only
targeting some simd and similar flags, and other flags will be kept
(albeit it does strip LTO unless Qt >=6.10 + >=gcc-15.2 or clang-any
where lto bugs no longer been occurring).

Bug: https://bugs.gentoo.org/966289
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>