dev-util/hip: do not include stdlibc++ file in HIPRTC header
When any hip header included with -D__HIPCC_RTC__, it means that
compiler was launched with -E -P to produce a preprocessed header.
In this case not a single libstdc++ header can be used, otherwise it
causes function redefinition on next attempt to include the same header
elsewhere.

Closes: https://bugs.gentoo.org/966188
Signed-off-by: Sv. Lockal <lockalsash@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/44725
Signed-off-by: Sam James <sam@gentoo.org>