distutils-r1.eclass: Add DISTUTILS_CONFIG_SETTINGS_JSON
Add DISTUTILS_CONFIG_SETTINGS_JSON that can be used to specify the JSON
data for config_settings directly.  For the time being, this is only
supported for standalone backends, since we can never know what
arguments they expect and we do not want the value to conflict with any
future defaults for known backends.

This is going to be used by the next version of x11-wm/qtile to control
building the Wayland support (currently handled by sed in the backend
code).

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Part-of: https://github.com/gentoo/gentoo/pull/44745
Closes: https://github.com/gentoo/gentoo/pull/44745
Signed-off-by: Michał Górny <mgorny@gentoo.org>