app-emacs/geiser-guile: do not inherit guile-single
guile-single was only used to overwrite the default value of
geiser-guile-binary on linux systems, this prevents the user from
controlling the version of guile used by geiser via eselect-guile or by
some other path manipulation, which means they would need to install
their own copy of geiser-guile as a workaround.

The ebuild is updated to RDEPEND on dev-scheme/guile without specifying
which slot should be used.

Signed-off-by: Tom Gillespie <tgbugs@gmail.com>