mail-client/claws-mail: depend on all possible libs for notification
USE=notification has three possible implementations/backends,
unfortunately they can not be controlled seperately and are
automagically enabled when installed on the build host.
In order to work around that we have to DEPEND on all three once
enabling the plugin.
Inttroduce NOTIFICATIONDEPEND so we can leave a note to related bugs.

Closes: https://bugs.gentoo.org/952993
Signed-off-by: Henning Schild <henning@hennsch.de>
Part-of: https://github.com/gentoo/gentoo/pull/44797
Closes: https://github.com/gentoo/gentoo/pull/44797
Signed-off-by: Sam James <sam@gentoo.org>