games-strategy/megaglest: Fix build w/ cmake-4
Upstream git master is already at 3.15 and 3.10 (Windows or else)
respectively, the little extra step should be fine. Maybe pass the info
on to upstream that it really is fine to depend on a CMake version released
in 2019.

While a 3.14 release appears to be in the (slow) works, we don't
know how long it will really take them.

Closes: https://bugs.gentoo.org/964656
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>