x11-misc/xdotool: backport WIP Fn keys fix for xkeyboard-config-2.46
Would have preferred to make the fix ~testing only given upstream has
not merged this yet and is still seeking feedback on whether this works
right, not to mention that this is a rebase applied to the (4 years old)
release that may or may not be missing something -- but 2.46 is already
stable and given users often use Fn keys in automation, this may break
several users' workflow.

For users experiencing issues, will keep the old unpatched -r1 in case
until a proper release. May mask -r2 if turns out it has severe issues.

Closes: https://bugs.gentoo.org/966686
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>