net-im/telegram-desktop: Remove libc++ suggestion
This message was put into place due to the frequent issues this package
has with the clang/libstdc++ configuration. It came to be from a
misunderstanding about the support status of the libc++ configuration.

Right now, this configuration is working. In the future, if it seems a
message like this is necessary again, it will only suggest building with
gcc.

Closes: https://bugs.gentoo.org/966774
Closes: https://github.com/gentoo/gentoo/pull/44818
Signed-off-by: Esteve Varela Colominas <esteve.varela@gmail.com>
Signed-off-by: Yixun Lan <dlan@gentoo.org>