www-client/*: update opera stable packaging
Opera decided to change the name of directories in stable.

It actually more closely matches -beta and -dev now, but in order to
ensure that these changes propagate as versions are promoted to stable
all packages get an update.

While we're there, drop the legacy `rpm` code - I've never seen a
release that _wasn't_ available as a .deb, and it's in the history if we
ever need it.

No revbumps as packages were either broken (stable) or have not
meaningfully changed.

Closes: https://bugs.gentoo.org/966737
Signed-off-by: Matt Jolly <kangie@gentoo.org>