media-libs/assimp: add some test excludes for USE="-collada"
The upstream testsuite does NOT implement filtering for tests that test
options which have been disabled.

With this change, we are down to one remaining failure:
- utIssues.OpacityBugWhenExporting_727

It passes with USE=collada but I cannot personally tell whether that
means the test relies on said feature or whether something else is
wrong, so I am not touching it for now. The test excludes I did add, are
*overtly* "Error: No suitable reader found for the file format xxxxx".

Bug: https://bugs.gentoo.org/962559
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>