net-misc/icaclient: switch to ffmpeg-compat:6
Not the maintainer nor know this application, but blocking ffmpeg-7
is rather nasty so going ahead with changing this.

Normally it'd use its bundled libav* inside ${ICAROOT}/lib but the
ebuild does `doexe *.so` which excludes libav's *.so.{60,58}.

Arguably using bundled would be fair if it's going to use some ancient
ffmpeg and potentially block future cleanups, but not going to impose
this change myself.

Given ffmpeg-compat:6 is not in LDPATH to avoid conflicts/mismatch, do
symlinks inside that directory which matches how it'd normally be used.

Not tried runtime but it seems to find them fine:

$ ldd libctxVideoDecoder.so
	libavcodec.so.60 => /opt/Citrix/ICAClient/lib/libavcodec.so.60
	libavutil.so.58 => /opt/Citrix/ICAClient/lib/libavutil.so.58

Closes: https://bugs.gentoo.org/967025
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>