sys-kernel/udev-hid-bpf: add more config options from kernel selftests
Extracted from tools/testing/selftests/hid/config in linux.git.

Some of the more device-specific ones of course may not be needed but
we don't have USE-per-quirk and I don't really want to add that, as upstream
have no way of doing that (they have a filter for install, not for building;
can use INSTALL_MASK for doing that or MYMESONARGS to set the install option).

Bug: https://bugs.gentoo.org/961038
Signed-off-by: Sam James <sam@gentoo.org>