media-video/popcorntime-bin: update src_uri and unbundle nwjs
The checksum at the new url is different, but the difference is only
in libatomic.so, which we never installed in the first place.

While digging into this I decided to unbundle nwjs completely which
cleans this up a lot and makes it possible to potentially push the
version of nwjs up. I tried version 100 but this did not work for
me, so pinning to the version that is bundled upstream for the time
being.

Closes: https://bugs.gentoo.org/963350
Closes: https://bugs.gentoo.org/927409
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>