net-analyzer/wireshark: make PKCS support explicit
Wireshark automagically enables PKCS support if it finds corresponding
symbols in gnutls at build time. This dependency is not recorded in
the ebuild's RDEPEND for gnutls, which means changes in gnutls might
break its ABI and consequently Wireshark.
Fix this by adding USE=pkcs11 support in the ebuild, forwarding this
to the Wireshark cmake build and depending on gnutls[pkcs11] when
required.

Closes: https://bugs.gentoo.org/967060
Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>
Part-of: https://github.com/gentoo/gentoo/pull/44894
Closes: https://github.com/gentoo/gentoo/pull/44894
Signed-off-by: Sam James <sam@gentoo.org>