x11-drivers/nvidia-drivers: fix build with ZONE_DEVICE + kernel 6.18
When marking as 6.18 compatible, I missed this USE=kernel-open build
failure given it only happens when CONFIG_ZONE_DEVICE is enabled and
the kernel I tested did not (option requires memory hotplugging and
hotremove support to be enabled, not enabled by defconfig).

Normally we do not backport kernel fixes, but in this case the only
fixed version is currently the 590.x (unkeyworded) beta and given
6.18 is due to become LTS I'd rather get this fixed now (possible
that new 0/vulkan releases will not include the fix too).

Only 580.x is affected given older branches did not use this function.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>