dev-cpp/abseil-cpp: build with c++17
Build with C++17 for now to avoid ABI hell with std::*_ordering with
C++20. If abseil is built with a compiler that defaults to C++20, like
GCC 16, then reverse dependencies start to fail unless they're also
built with C++20. Just use C++17 to have the status quo for now to
unblock more GCC testing.

[sam: Wrote the commit message.]

Closes: https://bugs.gentoo.org/966846
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>