net-misc/yt-dlp: relax yt-dlp-ejs dependency
Missed this when doing the bump, but just noticed that it no longer
hard errors if the patch version does not match. It still checks
the sha512sum of the .js files but they haven't changed since 0.3.0
and in theory won't until 0.4.x.

This should especially make things nicer for the live ebuild while
a yt-dlp-ejs update is pending. Not particularly important for the
release version though (not worth a revbump).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>