dev-build/cmake: add 4.2.1
I've dropped the BLAS patch as it no longer applies and it seems like a
good opportunity to cut it loose, now we have flexiblas (+ see the discussion
on the bug).

If it were to come back, it should be made opt-in and just enabled by
cmake.eclass, not to interfere with local development.

Closes: https://bugs.gentoo.org/736547
Signed-off-by: Sam James <sam@gentoo.org>