net-libs/czmq: remove nonfunctional USE=static-libs
Nobody has noticed for over 4 years that ever since the port to EAPI 8,
this didn't actually install any static libs. Reason is because EAPI 8
started passing --enable-static by default, and the ebuild implemented
the USE "incorrectly" by building useless stuff per EAPI 7 default and
then rm'ing it in src_install only if it existed. With the bump, it was
not built at all.

Given nobody can have ever relied on this in recent years, it is clear
nobody uses it, so no need to make it work.

Fixes: 1c4168493ca6cf243985a6f4efebc5986438295c
Fixes: 5afce7b17163b47ff8fb4df89c9f93aa3ef94a08
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>