acct-group/cuse: New group, v0
New group for /dev/cuse.  Unlike /dev/fuse, we don't want this device
to be world-accessible because of potential security concerns, starting
with the fact that CUSE is not widely tested, and ending with
CVE-2016-1233 which indicates that programs using CUSE can modify memory
of processes opening CUSE devices.

Bug: https://bugs.gentoo.org/967033
Signed-off-by: Michał Górny <mgorny@gentoo.org>
Part-of: https://codeberg.org/gentoo/gentoo/pulls/2
Signed-off-by: Michał Górny <mgorny@gentoo.org>