dev-python/mypy: Disable native-extensions by default, backported
As before, mypy with mypyc takes a *very* long time to build and has
questionable codegen approaches, zero parallelism, defies ccache, etc.
While some will surely desire it (I do!) some won't and it seems overall
more trouble than it's worth if you build your own from source. (This is
where the binhost comes in handy).

The latest iteration is discovering it consistently (for me) runs out of
memory on arm32, which again is not as surprising as it could be given
mypyc operates by "unity build"ing the entire codebase as a single `*.c`
file (not even one per module), resulting not just in long builds, but
builds where all costs are paid at the same time.

Sad to say, this default really needs to be given the boot.

Bug: https://bugs.gentoo.org/966484
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>