profiles: mask calibre 8.16
Regrettably, this non-maintainer upload is also a version that has lit a
firecracker under the community and attracted some ire. The situation is
in a state of mild flux, and more generally requires further
consideration to determine the correct course of action. Time which is
in somewhat limited supply for me right now.

I don't appreciate having my hand forced. Doing anything is picking a
side, and in this case I've effectively had someone pick my side for me.

There is no hurry here. There is a Qt 6.10 compat issue, but the
backports are trivial and were already diagnosed, and are needed anyway
to avoid ~arch holding up Qt.

Closes: https://bugs.gentoo.org/967338 (invalid)
Bug: https://bugs.gentoo.org/963607
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>