sys-apps/baselayout: Use branding.eclass and generate files in ebuild
It's easier to generate gentoo-release and os-release in the ebuild,
especially when using branding.eclass.

baselayout itself still generates these from the Makefile, but we can
drop that afterwards.

I've added VERSION to os-release with ${PV} to match VERSION_ID because
I've seen software (erroneously) fail without it, and it just makes
sense anyway.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>