app-emulation/wine-*: drop IUSE=xcomposite in live (enabled with X)
Was already default and feel there is no reason to ever disable this
with USE=X, and disabling seem to have recently been causing issues
with CEF / Steam and similar which may be hard to figure out for users.

Leaving off-by-default IUSE=xinerama alone given it is hardly needed
nowadays and can be considered deprecated, no reason to force users to
newly set USE=abi_x86_32 on an extra library over this.

Will backport the change to other versions probably after wine-11 is
stable (if do not forget) to minimize rebuilds. Likely will not revbump
all wine versions over even though it's a removed->enabled change.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>