dev-python/pypdf: add test dep on jbig2dec
It is an optional feature, and the tests can skip that test if
uninstalled. But if jbig2dec is installed, the tests use its png support
unconditionally, and fail if built without that USE.

Given a choice between:

- `BDEPEND="test? ( !!media-libs/jbig2dec[-png] )"`
- `BDEPEND="test? ( media-libs/jbig2dec[png] )"`
- src_test checking for and disabling the test

simply... enforcing decent test coverage is pretty simple and easy so I
can't think of any reason not to do it. It's a  simple package to depend
on. ;)

Bug: https://github.com/py-pdf/pypdf/issues/3550
Closes: https://bugs.gentoo.org/959843
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>