profiles/prefix/darwin/macos: drop 10.5 (Leopard) support
A bit overdue, however today's reality is that it will not be possible
without heavy patching and modifications to bootstrap on old 10.5
systems.  The fact that hardware running this (PPC G5s mostly) is dying,
if it hasn't died yet, and that running this under virtualisation seems
impossible also makes it unlikely there is a need, desire or means to
support this OS.

This means the end of an era, as the ppc-macos keyword becomes obsolete.
It served its purpose for many years, and was the first Gentoo for Mac
OS X keyword, as well as for the successor Gentoo Prefix project.  Its
popularity has topped the other keywords for a decade, but now it must go
to the history section of Gentoo, and Prefix in particular.

Bye bye old friend.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>