sys-libs/glibc: ignore CFLAGS+LDFLAGS in do_run_test for binpkgs
With binpkgs, CFLAGS and LDFLAGS will come from the builder environment
and may not be compatible with an old glibc on the system before a new
one is merged.

In the case reported to me by Kerin, an old glibc was installed w/o
DT_RELR support, and LDFLAGS on the binhost builder had -Wl,-z,pack-relative-relocs,
so the built test binary (which would run against the system's old glibc)
couldn't be executed, meaning the "syscall test" failed.

Set CFLAGS to just -O2 and blank LDFLAGS to avoid this.

It did also wrongly diagnose the failure because the function both
compiles + runs and the error message can't distinguish between those. If
we fail to build it, things are likely fine, there's only a problem if we
fail to run it. I haven't addressed that here.

Bug: https://bugs.gentoo.org/324685
Signed-off-by: Sam James <sam@gentoo.org>