dev-libs/quazip: stabilize arm64
It works as expected. Not much to say. I did initially stabilize riscv
but realized a bunch of deps aren't stabilized for that yet. So I'm
going to test those first when I return to my RISC-V mancave, then
stabilize if everything works as expected.

Note that I intentionally did
not mark arm as stable; not only do I not own a single ARMv7 device not
named the Nintendo 3DS (and good luck running a Qt app on there), but Qt
apparently has some issues on it anyways. This can be changed if you
wish.

btw, why is 1.5 still marked as unstable? It's been stable for a while.

Signed-off-by: crueter <crueter@eden-emu.dev>