dev-cpp/simpleini: keyword ~arm64
It works blah blah blah how many times do I need to write this. It also
works properly on riscv, but I haven't done enough tests to be sure.
Speaking of sure, are we sure we can't stabilize 4.25? I mean it works
as expected everywhere.

Note that I didn't keyword 4.22-r1 because I didn't test
it.

Signed-off-by: crueter <crueter@eden-emu.dev>