dev-debug/gdb: add 17.1
We could write up the new 'disable formats and functionality' configure
options but it's not clear to me there's significant benefit in doing so,
or how we'd divide that up, so leave it, at least for now and possibly
forever.

> New option --disable-gdb-compile to disable support for the compile subsystem.
> New option --disable-gdb-dwarf-support to disable support for DWARF debug information in GDB.
> New option --disable-gdb-mdebug-support to disable support for mdebug/ecoff debug information.

The main motivation upstream I understand was because of (debatable) CVEs
in lesser-used formats.

Signed-off-by: Sam James <sam@gentoo.org>