x11-misc/gammastep: Resolve broken use flag "appindicator" by merging into "gtk"
Flag "appindicator" was not activating compilation of command
"gammastep-indicator". That was done by flag "gtk". When deciding which of
the two flags to dissolve, I decided to keep the one that was providing the
command but failed to pull in all needed dependencies.
The issue was inherited from package x11-misc/redshift originally. I will
fix it there as well in a second.

Signed-off-by: Sebastian Pipping <sping@gentoo.org>