x11-misc/redshift: Resolve broken use flag "appindicator" by merging into "gtk"
Flag "appindicator" was not activating compilation of command "redshift-gtk".
That was done by flag "gtk". When deciding which of the two flags to dissolve,
I decided to keep the one that was providing the command but failed to pull in
all needed dependencies.
The very same issue has also been fixed for package x11-misc/gammastep in the
previous commit.

Signed-off-by: Sebastian Pipping <sping@gentoo.org>