toolchain-funcs.eclass: fix pattern for detecting ld.bfd
See the test case below:

> out=$(echo -e "abcdef\n123456"); [[ ${out} =~ .*^"123456".* ]] && echo "matched" || echo "not matched"

This prints "matched" when using glibc, but "not matched" when using musl.

The root cause is that using ^ in the middle of the regex is undefined behavior.
Replace it with (^|$'\n'), i.e. match either the beginning of the string or a
position immediately following a newline.

Signed-off-by: Z. Liu <zhixu.liu@gmail.com>