media-libs/subrandr: new package, add 1.0.0
New optional dependency of mpv-9999 to display some non-ASS subtitle
formats (currently mostly formats used by the web+youtube).

Unfortunately does not use cargo-c but rather builds some rust-written
builder (xtask) and then uses it to build+install and generate the
.pc file among other things.

Have not tested cross-compilation and there is good chances that it
will need fixing (aka, maybe either xtask is not built for CBUILD or
subrandr isn't built for CHOST and will need to pass some --target
w/ $(rust_abi) which xtask does support like cargo-c).

Not doing multilib for now, but we may end up wanting it in the future
if e.g. ffmpeg adds support... albeit it's unlikely old 32bit
applications will need this and we could leave support native-only.

Currently unknown if/when this will need a subslot, upstream claims
the C API should be stable unless packages set the SBR_ALLOW_UNSTABLE
macro. We *could* need it only on packages that set that if any.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>