media-libs/libcamera: new package, add 0.6.0
Important notes:

While libcamera support optional IPA module signing, this ebuild
enforces it. This approach simplify the combination of USE flags.

OpenSSL is the preferred crypto library because it is available by default on
most systems and GnuTLS is the fallback. However, even if GnuTLS is selected,
OpenSSL is required at build time.

IPA modules are FAT binaries. Trying to strip those binaries invalidate the
computed signatures. Trying to re-sign those modules introduced some nuances
on bindist packaging and probe to be difficult to maintain. This will be
addressed later.

Future work:
  - Support for stripping IPA signing modules. We would like to
    re-evaluate how to re-sign IPA modules via patching ELF binaries directly.
  - Support for Python bindings (pycamera).
  - Support for documentation generation.

Bug: https://bugs.gentoo.org/867670
Signed-off-by: Kellerman Rivero Suarez <krsloco@gmail.com>