sci-libs/blas-lapack-aux-wrapper: fix forcing netlib for tests
FLEXIBLAS=trigger-warning doesn't work because the fallback isn't
netlib, it is "what would have been used if the variable is unset".
Explicitly set netlib.

Getting log info regarding flexiblas being used can be done by setting
FLEXIBLAS_VERBOSE. It is much more chatty, far more than one line gets
into the logfile, but that's okay.

Fixes tests on arm, where openblas (the USE default of flexiblas!) seems
to just not work as expected. This should not be allowed to affect the
testsuite for blas-lapack-aux-wrapper, which is testing if the wrapper
works, not if openblas does.

Bug: https://bugs.gentoo.org/967928
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>