dev-python/primecountpy: block primecount >= 8.0
This won't build against primecount-8.0 yet, but a new release should be
coming soon.

I've dropped the old stable ebuild and stabilized the -r1 instead. We
would have needed to edit the dependencies on the stable version anyway
to prevent the package manager from thinking that a downgrade could
solve the conflict, and then the RDEPEND change would have needed a
new revision... in the end, this seems like the least bad option.

Closes: https://bugs.gentoo.org/967990
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>