acct-group.eclass: add support for ACCT_GROUP_<UPPERCASE_GROUPNAME>_USERS_ADD
There are cases, when we want the package manager to add users
which aren't managed by the package manager, to groups managed
by the PM. A usecase we found for it is adding normal users to
acct-group/jobserver.

When used, we print a warning that support will not be provided.
We also support non existing users (will print warning, but
continue without handling them).

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>