nginx-module.eclass: Use response file in ngx_mod_append_libs()
Previously, the arguments were appended into the patched-in environment
variable that would feed its contents into the build system of NGINX,
which caused many funny issues that can be discovered, for example, by
looking this commit's diff.

This commit instead makes use of the awesome response files, which allow
putting arguments in a file and have the compiler interpret them
directly without 4 layers of eval and a very helpful Make
interpretation. Thank you.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>