games-arcade/supertux: add 0.7.0_beta1
There is one pretty gross bit here, which is the
SUPERTUX_{PACKAGE_VERSION,VERSION_STRING} variables here. This is
because the beta actually seems to have a little logic issue in the
git tag revision cmake, which essentially means that it fails during
an attempt to look for the .git folder (i'm serious). As a workaround,
we just hardcode the package version & version string to the our
PV. This is pretty much how it'd look in a real release anyway. It'll
certainly be fixed upstream obviously.

New use flags for opengl are here as well, as that's now optional
(always has been), and gtest was dropped because we just use ctest
now.

Signed-off-by: Swagtoy <me@ow.swag.toys>